vertex-ai/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0